api/serializers.py
from django.test import TestCase

# Create your tests here.
from rest_framework import serializers
from .models import Task

class TaskSerializer(serializers.ModelSerializer):
	class Meta:
		model = Task
		fields ='__all__'
